+++ js/data.js
@@ -1,0 +1,36 @@
+const PHOTO_DESCRIPTIONS = [
+  'Закат на берегу моря',
+  'Уютное утро с чашкой кофе',
+  'Прогулка по осеннему парку',
+  'Вид на город с высоты',
+  'Сочное летнее манго',
+  'Ночной свет звезд',
+  'Солнечные лучи в лесу',
+  'Тихая гавань у моря',
+  'Уличное искусство в центре города',
+  'Вечер с любимой книгой',
+];
+
+const PHOTO_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES_USER = [
+  'Александр',
+  'Мария',
+  'Дмитрий',
+  'Анна',
+  'Иван',
+  'Екатерина',
+  'Сергей',
+  'Ольга',
+  'Андрей',
+  'Наталья'
+];
+
+export {PHOTO_DESCRIPTIONS, PHOTO_COMMENTS, NAMES_USER};
